backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
sqlalchemy
pydantic-settings
pyjwt
pwdlib
httpx
